Rename `ServerComponentResponse` to `HydrogenResponse`
