+++ packages/hydrogen/src/utilities/log/__tests__/log-cache-header.test.ts
@@ -7,7 +7,7 @@ import {
   setLoggerOptions,
 } from '../index';
 import {HydrogenRequest} from '../../../framework/HydrogenRequest.server';
-import {ServerComponentResponse} from '../../../framework/Hydration/ServerComponentResponse.server';
+import {HydrogenResponse} from '../../../framework/HydrogenResponse.server';
 
 let mockLogger: jest.Mocked<Logger>;
 
@@ -45,7 +45,7 @@ describe('cache header log', () => {
     } as unknown as HydrogenRequest;
     const response = {
       cacheControlHeader: 'public, max-age=1, stale-while-revalidate=9',
-    } as ServerComponentResponse;
+    } as HydrogenResponse;
 
     logCacheControlHeaders('str', request, response);
 
@@ -69,7 +69,7 @@ describe('cache header log', () => {
     } as unknown as HydrogenRequest;
     const response = {
       cacheControlHeader: 'public, max-age=1, stale-while-revalidate=9',
-    } as ServerComponentResponse;
+    } as HydrogenResponse;
 
     logCacheControlHeaders('rsc', request, response);
 
@@ -93,7 +93,7 @@ describe('cache header log', () => {
     } as unknown as HydrogenRequest;
     const response = {
       cacheControlHeader: 'public, max-age=1, stale-while-revalidate=9',
-    } as ServerComponentResponse;
+    } as HydrogenResponse;
 
     collectQueryCacheControlHeaders(
       request,
@@ -126,7 +126,7 @@ describe('cache header log', () => {
     } as unknown as HydrogenRequest;
     const response = {
       cacheControlHeader: 'public, max-age=1, stale-while-revalidate=9',
-    } as ServerComponentResponse;
+    } as HydrogenResponse;
 
     collectQueryCacheControlHeaders(
       request,
